Scoring
Function scoring takes in two parameters the training data and testing
data and gives us the confidence score and the number of matching
attributes for each feature of the training data and outputs the result
in a text file - output.txt

As there was no test data I put in one random test data. Comment Line
100 when the test results are available(no need to change anything in
the code as the variables are set accordingly).
Also, the paths have been changed to sync with MAC, windows users need
to change it accordingly.